+++ src/boilerplate/rendering/textures/Texture2d.java
@@ -15,7 +15,8 @@ import java.util.HashMap;
 import java.util.Map;
 
 import static org.lwjgl.opengl.GL11.*;
-import static org.lwjgl.opengl.GL13.*;
+import static org.lwjgl.opengl.GL13.GL_TEXTURE0;
+import static org.lwjgl.opengl.GL13.glActiveTexture;
 import static org.lwjgl.opengl.GL45.glBindTextureUnit;
 
 public class Texture2d extends Texture {
@@ -100,7 +101,7 @@ public class Texture2d extends Texture {
 
         if (boundSlots.containsKey(slot)) {
             Logging.warn("Overriding already set texture slot '%s'", slot);
-            boundSlots.remove((Integer) slot);
+            boundSlots.remove(slot);
         }
 
         glBindTextureUnit(slot, textureId);
